+++ src/modules/starting-fetch/starting-fetch.js
@@ -1,11 +1,6 @@
-import { globalGenres } from './globalGenres';
-import { refs } from '../refs';
 import fetchApi from '../api-service';
 import { apiRefs } from '../api-service';
-const BASE_URL = 'https://api.themoviedb.org/3';
-const BASE_IMAGE_URL = 'https://image.tmdb.org/t/p/';
-const API_KEY = 'cb1bcc244723619ea7f2217b5a84ccd8';
-const imageSize = 'w500';
+import renderMoviesList from '../html-render';
 
 const startingFetch = async () => {
   const response = await fetchApi(apiRefs.TRENDING);
@@ -26,53 +21,5 @@ const getStartingArray = async () => {
 };
 
 getStartingArray().then(() => {
-  renderStartingMoviesList(startingArr);
+  renderMoviesList(startingArr);
 });
-
-function renderStartingMoviesList(moviesArr) {
-  const markup = moviesArr
-    .map(movie => {
-      const {
-        title,
-        name,
-        poster_path,
-        id,
-        release_date,
-        first_air_date = '',
-        genre_ids = [],
-      } = movie;
-      let movieYear = release_date
-        ? getMovieYear(release_date)
-        : getMovieYear(first_air_date);
-      let movieName = title ? title : name;
-      const movieGenres = getMovieGenres(genre_ids, globalGenres);
-      const fullImageUrl = `${BASE_IMAGE_URL}${imageSize}${poster_path}`;
-      return `<li class="movie" data-id=${id}>
-<img src="${fullImageUrl}" width="395"/>
-<p class="movie__title">${movieName}</p>
-<p class="movie__genres">${movieGenres}</p>
-<p class="movie__year">${movieYear}</p>
-</li>
-`;
-    })
-    .join('');
-  refs.galleryListRef.innerHTML = markup;
-}
-
-function getMovieYear(releasedate) {
-  return releasedate.slice(0, 4);
-}
-
-function getMovieGenres(genreIdsArray, genres) {
-  const genreNames = [];
-  genreIdsArray.forEach(id => {
-    if (genreNames.length >= 2) {
-      return;
-    }
-    const genre = genres.find(genreObj => genreObj.id === id);
-    if (genre) {
-      genreNames.push(genre.name);
-    }
-  });
-  return genreNames.join(', ');
-}
